# Use unaligned 32-bit and 64-bit compare in longest_match