- Use unaligned 32-bit and 64-bit compare based on best match length when searching for matches.
- Move TRIGGER_LEVEL to match_tpl.h since it is only used in longest match.
- Use early return inside match loops instead of cont variable.
- Added back two variable check for platforms that don't supported unaligned access.