@nmoinvaz 2020-07-28
```
 ## HEAD
 
 Runs: 32
 Levels: 0-9
 Trimworst: 8

 Level   Comp   Comptime min/avg/max                          Compressed size
 0     100.015%    0.291/0.326/0.340                          211979883
 1      47.684%    1.115/1.222/1.262                          101065293
 2      35.540%    1.956/2.122/2.192                          75325919
 3      34.226%    2.311/2.499/2.555                          72541436
 4      32.953%    2.752/3.056/3.128                          69843427
 5      32.665%    3.032/3.262/3.341                          69233255
 6      32.515%    3.474/3.813/3.907                          68914325
 7      32.257%    4.752/5.133/5.287                          68367431
 8      32.172%    7.296/7.933/8.150                          68186831
 9      32.156% 10.326/10.972/11.180                          68154021

 avg1   41.218%                4.034
 avg2   45.798%                4.482
 tot                         968.156

## PR

  Level   Comp   Comptime min/avg/max                          Compressed size
 0     100.015%    0.290/0.317/0.342                          211979883
 1      47.684%    1.079/1.176/1.291                          101065293
 2      35.540%    1.850/2.020/2.191                          75326374
 3      34.203%    2.271/2.455/2.634                          72492718
 4      32.936%    2.762/2.980/3.198                          69806244
 5      32.665%    2.985/3.207/3.461                          69233255
 6      32.515%    3.487/3.741/3.988                          68914325
 7      32.257%    4.601/4.929/5.233                          68367431
 8      32.172%    7.095/7.577/8.090                          68186831
 9      32.156%  9.825/10.485/11.153                          68154021

 avg1   41.214%                3.889
 avg2   45.794%                4.321
 tot                         933.285
```