@nmoinvaz 2020-07-29
New benchmarks:

## HEAD 
```
 Runs: 32
 Levels: 0-9
 Trimworst: 8

 Level   Comp   Comptime min/avg/max                          Compressed size
 0     100.015%    0.290/0.308/0.320                          211979883
 1      47.684%    1.034/1.047/1.057                          101065293
 2      35.540%    1.812/1.844/1.870                          75325919
 3      34.226%    2.160/2.189/2.211                          72541436
 4      32.953%    2.619/2.640/2.666                          69843427
 5      32.665%    2.840/2.858/2.874                          69233255
 6      32.515%    3.323/3.349/3.368                          68914325
 7      32.257%    4.515/4.537/4.562                          68367431
 8      32.172%    6.977/7.007/7.035                          68186831
 9      32.156%    9.639/9.692/9.728                          68154021

 avg1   41.218%                3.547
 avg2   45.798%                3.941
 tot                         851.258

 Runs: 32
 Levels: 0-9
 Trimworst: 8

 Level   Comp   Comptime min/avg/max                          Compressed size
 0     100.015%    0.290/0.296/0.304                          211979883
 1      47.684%    1.024/1.050/1.103                          101065293
 2      35.540%    1.804/1.848/1.875                          75325919
 3      34.226%    2.149/2.190/2.238                          72541436
 4      32.953%    2.617/2.662/2.731                          69843427
 5      32.665%    2.811/2.865/2.934                          69233255
 6      32.515%    3.323/3.370/3.442                          68914325
 7      32.257%    4.512/4.562/4.678                          68367431
 8      32.172%    6.973/7.074/7.258                          68186831
 9      32.156%   9.651/9.755/10.007                          68154021

 avg1   41.218%                3.567
 avg2   45.798%                3.964
 tot                         856.146
```
## COMMIT 0
 ```
 Runs: 32
 Levels: 0-9
 Trimworst: 8

 Level   Comp   Comptime min/avg/max                          Compressed size
 0     100.015%    0.330/0.338/0.343                          211979883
 1      47.684%    1.030/1.045/1.061                          101065293
 2      35.540%    1.793/1.820/1.848                          75326374
 3      34.203%    2.181/2.225/2.245                          72492718
 4      32.936%    2.696/2.715/2.734                          69806244
 5      32.665%    2.906/2.927/2.938                          69233255
 6      32.515%    3.367/3.397/3.415                          68914325
 7      32.257%    4.476/4.493/4.514                          68367431
 8      32.172%    6.887/6.911/6.935                          68186831
 9      32.156%    9.520/9.541/9.579                          68154021

 avg1   41.214%                3.541
 avg2   45.794%                3.935
 tot                         849.883

 Runs: 32
 Levels: 0-9
 Trimworst: 8

 Level   Comp   Comptime min/avg/max                          Compressed size
 0     100.015%    0.289/0.294/0.299                          211979883
 1      47.684%    1.027/1.038/1.051                          101065293
 2      35.540%    1.785/1.807/1.840                          75326374
 3      34.203%    2.179/2.208/2.241                          72492718
 4      32.936%    2.675/2.700/2.714                          69806244
 5      32.665%    2.882/2.902/2.919                          69233255
 6      32.515%    3.351/3.380/3.401                          68914325
 7      32.257%    4.462/4.489/4.512                          68367431
 8      32.172%    6.854/6.891/6.918                          68186831
 9      32.156%    9.505/9.534/9.557                          68154021

 avg1   41.214%                3.524
 avg2   45.794%                3.916
 tot                         845.842
 ```
## COMMIT 1
```
 Runs: 32
 Levels: 0-9
 Trimworst: 8

 Level   Comp   Comptime min/avg/max                          Compressed size
 0     100.015%    0.288/0.297/0.304                          211979883
 1      47.684%    1.030/1.043/1.055                          101065293
 2      35.540%    1.758/1.779/1.794                          75326374
 3      34.203%    2.141/2.176/2.203                          72492718
 4      32.936%    2.631/2.655/2.681                          69806244
 5      32.665%    2.836/2.867/2.888                          69233255
 6      32.515%    3.315/3.334/3.348                          68914325
 7      32.257%    4.407/4.432/4.458                          68367431
 8      32.172%    6.812/6.842/6.863                          68186831
 9      32.156%    9.455/9.494/9.522                          68154021

 avg1   41.214%                3.492
 avg2   45.794%                3.880
 tot                         838.066

 Runs: 32
 Levels: 0-9
 Trimworst: 8

 Level   Comp   Comptime min/avg/max                          Compressed size
 0     100.015%    0.293/0.299/0.309                          211979883
 1      47.684%    1.024/1.036/1.085                          101065293
 2      35.540%    1.756/1.789/1.837                          75326374
 3      34.203%    2.136/2.177/2.210                          72492718
 4      32.936%    2.626/2.662/2.693                          69806244
 5      32.665%    2.833/2.859/2.886                          69233255
 6      32.515%    3.299/3.333/3.370                          68914325
 7      32.257%    4.389/4.435/4.501                          68367431
 8      32.172%    6.785/6.847/6.885                          68186831
 9      32.156%    9.442/9.490/9.546                          68154021

 avg1   41.214%                3.493
 avg2   45.794%                3.881
 tot                         838.258
```