@Dead2 2020-08-15
This looks very interesting. I see that the first commit changes the output size of compression levels 2-4 in my tests, do we know what specifically caused that change and that it is safe? In my tests it results in decreased size for all three levels, but I see that one of your tests ended up bigger. It looks like this change is only in the `level < TRIGGER_LEVEL` case.

My tests confirm that this results in a speedup, especially in the higher levels. In my tests level 9 is ~2.9% faster, and that is quite significant. For levels 1-2 the speed seems to be on par with before, with the change being well within the error margins.

Compiled code size increases quite a bit, but I am not too worried about that.

We should try to understand the nature of the output change and decide whether it is truly for the better, if it is then this looks very good indeed.

## Deflatebench config
 Runs: 70
 Levels: 1-9
 Trimworst: 40

Activated multiple generated file mode. Source: silesia-small.tar
Level 1: /tmp/deflatebench-1.tmp 270.1 MiB / 283,253,760 B
Level 2: /tmp/deflatebench-2.tmp 135.1 MiB / 141,626,880 B
Level 3: /tmp/deflatebench-3.tmp 105.1 MiB / 110,154,240 B
Level 4: /tmp/deflatebench-4.tmp 90.0 MiB / 94,417,920 B
Level 5: /tmp/deflatebench-5.tmp 90.0 MiB / 94,417,920 B
Level 6: /tmp/deflatebench-6.tmp 75.0 MiB / 78,681,600 B
Level 7: /tmp/deflatebench-7.tmp 60.0 MiB / 62,945,280 B
Level 8: /tmp/deflatebench-8.tmp 45.0 MiB / 47,208,960 B
Level 9: /tmp/deflatebench-9.tmp 45.0 MiB / 47,208,960 B


## Baseline 8832d7db7241194fa68509c96c092f3cf527ccce
```
   text    data     bss     dec     hex filename
  93335    1272      32   94639   171af libz-ng.so.1

 Level   Comp   Comptime min/avg/max  Decomptime min/avg/max  Compressed size
 1      57.870%    1.863/1.880/1.891       1.181/1.215/1.234  163918126
 2      43.903%    1.825/1.854/1.863       0.614/0.631/0.640  62178636
 3      42.510%    1.700/1.724/1.734       0.460/0.476/0.483  46826710
 4      41.490%    1.681/1.699/1.707       0.387/0.399/0.407  39174432
 5      41.215%    1.825/1.849/1.857       0.372/0.396/0.405  38913920
 6      41.038%    1.834/1.845/1.851       0.323/0.332/0.337  32289128
 7      40.784%    1.998/2.014/2.021       0.253/0.261/0.269  25671760
 8      40.707%    1.935/1.949/1.955       0.185/0.194/0.199  19217358
 9      40.696%    2.154/2.164/2.170       0.185/0.195/0.201  19212387

 avg1   43.357%                1.887                   0.456
 tot                         509.364                 122.992
```

## PR #707 95eb3415c46aac9c79b6c5b2cb96b602a99f2ae7
```
   text    data     bss     dec     hex filename
  94871    1272      32   96175   177af libz-ng.so.1

 Level   Comp   Comptime min/avg/max  Decomptime min/avg/max  Compressed size
 1      57.870%    1.852/1.876/1.888       1.176/1.212/1.231  163918126
 2      43.900%    1.845/1.857/1.868       0.615/0.630/0.638  62174655
 3      42.491%    1.751/1.767/1.776       0.463/0.477/0.486  46805381
 4      41.474%    1.712/1.726/1.735       0.378/0.397/0.405  39158901
 5      41.215%    1.855/1.870/1.880       0.380/0.395/0.406  38913920
 6      41.038%    1.814/1.841/1.851       0.315/0.331/0.338  32289128
 7      40.784%    1.954/1.966/1.972       0.253/0.263/0.269  25671760
 8      40.707%    1.879/1.894/1.901       0.182/0.195/0.201  19217358
 9      40.696%    2.090/2.101/2.106       0.178/0.195/0.201  19212387

 avg1   43.353%                1.878                   0.455
 tot                         506.955                 122.858
```